style: indent requirements files in prep for make upgrade
The latest version of pip-tools inserts much more
indentation into requirements/*.txt files, which is
great for general readability, but makes the very next
'make upgrade' diff much harder to read.

For the sake of this PR's review, I manually
did some of the indentation that pip-tools will do,
such that the following 'make upgrade' commit is easier
to review.