build: un-constrain attrs and factoryboy to fix upgrades
Upgrade jobs are failing because we're constraining to a 2018
version of attrs. git-blame yields no insight on why we are
doing this; it might be incidental. Let's remove the constraint
and just see what happens.

Also, factoryboy was constrained due to a Docker interaction
issue that is now resolved, so let's remove that constriant too.