build: include sub-packages in wheel
This repository's setup.py was configured
such that only the root Python package (`blockstore`)
was included, not sub-packages (`blockstore.apps`, etc.).
This commit fixes that by using `find_packages(...)`.

A similar fix was made to the cookiecutter repository
here: https://github.com/openedx/edx-cookiecutters/pull/257

The implication of the misconfiguration was that
blockstore could be installed in *editable* mode,
but not as a built package:

    # Editable installation.
    # This worked, by chance.
    pip install -e git+https://github.com/openedx/blockstore.git@1.2.4#egg=blockstore==1.2.4
    python -c "from blockstore import apps"

    # Built installation.
    # This failed with "ImportError: cannot import name 'apps' from 'blockstore'"
    pip install "blockstore @ git+https://github.com/openedx/blockstore.git@1.2.4"
    python -c "from blockstore import apps"

If not fixed, this would become a problem when blockstore
is pushed to PyPI and installed as just:

    pip install openedx-blockstore