fix: allow bundles to be created via django admin
Previously, trying to create a bundle would lead to the
message:

    IntegrityError at /admin/bundles/bundle/add/
    (1048, "Column 'collection_id' cannot be null")

because `collection` was indeed a required field, but
it was marked as editable=False, so it one could not
actually set it when making a new bundle.

With this change, we set editable=True, and add
some logic to the admin site so that the
`collection` field can be set when a bundle
is first created, but not afterwards.

This will trigger a migration, but since `editable`
is purely a Django-level validation field, no
changes should be seen in the MySQL schema.